fix: rename function
